subsea-drone-autopilot/app/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
sse-starlette==2.1.0
databricks-sdk==0.30.0
psycopg2-binary==2.9.9
httpx==0.27.0
